+++ neo/framework/async/AsyncServer.cpp
@@ -1231,8 +1231,8 @@ void idAsyncServer::ProcessUnreliableClientMessage( int clientNum, const idBitMs
 		return;
 	}
 
-	acknowledgeSequence = msg.ReadLong();
-	clientGameInitId = msg.ReadLong();
+	acknowledgeSequence = msg.ReadInt();
+	clientGameInitId = msg.ReadInt();
 
 	// while loading a map the client may send empty messages to keep the connection alive
 	if ( clientGameInitId == GAME_INIT_ID_MAP_LOAD ) {
@@ -1263,7 +1263,7 @@ void idAsyncServer::ProcessUnreliableClientMessage( int clientNum, const idBitMs
 		return;
 	}
 
-	client.acknowledgeSnapshotSequence = msg.ReadLong();
+	client.acknowledgeSnapshotSequence = msg.ReadInt();
 
 	if ( client.clientState == SCS_CONNECTED ) {
 
@@ -1305,7 +1305,7 @@ void idAsyncServer::ProcessUnreliableClientMessage( int clientNum, const idBitMs
 			break;
 		}
 		case CLIENT_UNRELIABLE_MESSAGE_PINGRESPONSE: {
-			client.clientPing = realTime - msg.ReadLong();
+			client.clientPing = realTime - msg.ReadInt();
 			break;
 		}
 		case CLIENT_UNRELIABLE_MESSAGE_USERCMD: {
@@ -1313,7 +1313,7 @@ void idAsyncServer::ProcessUnreliableClientMessage( int clientNum, const idBitMs
 			client.clientPrediction = msg.ReadShort();
 
 			// read user commands
-			clientGameFrame = msg.ReadLong();
+			clientGameFrame = msg.ReadInt();
 			numUsercmds = msg.ReadByte();
 			for ( last = NULL, i = clientGameFrame - numUsercmds + 1; i <= clientGameFrame; i++ ) {
 				index = i & ( MAX_USERCMD_BACKUP - 1 );
@@ -1486,7 +1486,7 @@ void idAsyncServer::ProcessChallengeMessage( const netadr_t from, const idBitMsg
 	idBitMsg	outMsg;
 	byte		msgBuf[MAX_MESSAGE_SIZE];
 
-	clientId = msg.ReadLong();
+	clientId = msg.ReadInt();
 
 	oldest = 0;
 	oldestTime = 0x7fffffff;
@@ -1678,7 +1678,7 @@ void idAsyncServer::ProcessConnectMessage( const netadr_t from, const idBitMsg &
 	char		password[ 17 ];
 	int			i, ichallenge, islot, OS, numClients;
 
-	protocol = msg.ReadLong();
+	protocol = msg.ReadInt();
 	OS = msg.ReadShort();
 
 	// check the protocol version
@@ -1688,10 +1688,10 @@ void idAsyncServer::ProcessConnectMessage( const netadr_t from, const idBitMsg &
 		return;
 	}
 
-	clientDataChecksum = msg.ReadLong();
-	challenge = msg.ReadLong();
+	clientDataChecksum = msg.ReadInt();
+	challenge = msg.ReadInt();
 	clientId = msg.ReadShort();
-	clientRate = msg.ReadLong();
+	clientRate = msg.ReadInt();
 
 	// check the client data - only for non pure servers
 	if ( !sessLocal.mapSpawnData.serverInfo.GetInt( "si_pure" ) && clientDataChecksum != serverDataChecksum ) {
@@ -1892,7 +1892,7 @@ bool idAsyncServer::VerifyChecksumMessage( int clientNum, const netadr_t *from, 
 	// pak checksums, in a 0-terminated list
 	numChecksums = 0;
 	do {
-		i = msg.ReadLong( );
+		i = msg.ReadInt( );
 		checksums[ numChecksums++ ] = i;
 		// just to make sure a broken client doesn't crash us
 		if ( numChecksums >= MAX_PURE_PAKS ) {
@@ -1904,7 +1904,7 @@ bool idAsyncServer::VerifyChecksumMessage( int clientNum, const netadr_t *from, 
 	numChecksums--;
 
 	// code pak checksum
-	gamePakChecksum = msg.ReadLong( );
+	gamePakChecksum = msg.ReadInt( );
 
 	fileSystem->GetPureServerChecksums( serverChecksums, OS, &serverGamePakChecksum );
 	assert( serverChecksums[ 0 ] );
@@ -1939,7 +1939,7 @@ void idAsyncServer::ProcessPureMessage( const netadr_t from, const idBitMsg &msg
 	int		iclient, challenge, clientId;
 	idStr	reply;
 
-	challenge = msg.ReadLong();
+	challenge = msg.ReadInt();
 	clientId = msg.ReadShort();
 
 	if ( ( iclient = ValidateChallenge( from, challenge, clientId ) ) == -1 ) {
@@ -1971,7 +1971,7 @@ void idAsyncServer::ProcessReliablePure( int clientNum, const idBitMsg &msg ) {
 	byte		msgBuf[MAX_MESSAGE_SIZE];
 	int			clientGameInitId;
 
-	clientGameInitId = msg.ReadLong();
+	clientGameInitId = msg.ReadInt();
 	if ( clientGameInitId != gameInitId ) {
 		common->DPrintf( "client %d: ignoring reliable pure from an old gameInit (%d)\n", clientNum, clientGameInitId );
 		return;
@@ -2070,7 +2070,7 @@ void idAsyncServer::ProcessGetInfoMessage( const netadr_t from, const idBitMsg &
 
 	common->DPrintf( "Sending info response to %s\n", Sys_NetAdrToString( from ) );
 
-	challenge = msg.ReadLong();
+	challenge = msg.ReadInt();
 
 	outMsg.Init( msgBuf, sizeof( msgBuf ) );
 	outMsg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
@@ -2686,9 +2686,9 @@ void idAsyncServer::ProcessDownloadRequestMessage( const netadr_t from, const id
 	int			dlRequest;
 	int			voidSlots = 0;				// to count and verbose the right number of paks requested for downloads
 
-	challenge = msg.ReadLong();
+	challenge = msg.ReadInt();
 	clientId = msg.ReadShort();
-	dlRequest = msg.ReadLong();
+	dlRequest = msg.ReadInt();
 
 	if ( ( iclient = ValidateChallenge( from, challenge, clientId ) ) == -1 ) {
 		return;
@@ -2700,7 +2700,7 @@ void idAsyncServer::ProcessDownloadRequestMessage( const netadr_t from, const id
 	}
 
 	// the first token of the pak names list passed to the game will be empty if no game pak is requested
-	dlGamePak = msg.ReadLong();
+	dlGamePak = msg.ReadInt();
 	if ( dlGamePak ) {
 		if ( !( dlSize[ 0 ] = fileSystem->ValidateDownloadPakForChecksum( dlGamePak, pakbuf, true ) ) ) {
 			common->Warning( "client requested unknown game pak 0x%x", dlGamePak );
@@ -2715,7 +2715,7 @@ void idAsyncServer::ProcessDownloadRequestMessage( const netadr_t from, const id
 	numPaks = 1;
 
 	// read the checksums, build path names and pass that to the game code
-	dlPakChecksum = msg.ReadLong();
+	dlPakChecksum = msg.ReadInt();
 	while ( dlPakChecksum ) {
 		if ( !( dlSize[ numPaks ] = fileSystem->ValidateDownloadPakForChecksum( dlPakChecksum, pakbuf, false ) ) ) {
 			// we pass an empty token to the game so our list doesn't get offset
@@ -2725,7 +2725,7 @@ void idAsyncServer::ProcessDownloadRequestMessage( const netadr_t from, const id
 		}
 		pakNames.Append( pakbuf );
 		numPaks++;
-		dlPakChecksum = msg.ReadLong();
+		dlPakChecksum = msg.ReadInt();
 	}
 
 	for ( i = 0; i < pakNames.Num(); i++ ) {
